requirements.txt
streamlit==1.28.1
azure-ai-documentintelligence==1.0.0b1
azure-ai-formrecognizer==3.3.0
openai==1.99.9
python-dotenv==1.0.0
Pillow==10.0.1
json5==0.9.14
requests==2.31.0